+++ paymentsheet/src/test/java/com/stripe/android/paymentsheet/PaymentOptionsViewModelTest.kt
@@ -481,7 +481,7 @@ internal class PaymentOptionsViewModelTest {
         }
 
     @Test
-    fun `currentScreen is Form if payment methods is empty and supportedPaymentMethods contains one in automatic`() =
+    fun `currentScreen is AddFirstPaymentMethod if payment methods is empty and supportedPaymentMethods contains one in automatic`() =
         runTest {
             val viewModel = createViewModel(
                 args = PAYMENT_OPTION_CONTRACT_ARGS.updateState(
@@ -498,12 +498,12 @@ internal class PaymentOptionsViewModelTest {
             )
 
             viewModel.navigationHandler.currentScreen.test {
-                assertThat(awaitItem()).isInstanceOf<PaymentSheetScreen.VerticalModeForm>()
+                assertThat(awaitItem()).isInstanceOf<AddFirstPaymentMethod>()
             }
         }
 
     @Test
-    fun `currentScreen is VerticalMode if payment methods is not empty in automatic`() =
+    fun `currentScreen is SelectSavedPaymentMethods if payment methods is not empty in automatic`() =
         runTest {
             val viewModel = createViewModel(
                 args = PAYMENT_OPTION_CONTRACT_ARGS.updateState(
@@ -520,12 +520,12 @@ internal class PaymentOptionsViewModelTest {
             )
 
             viewModel.navigationHandler.currentScreen.test {
-                assertThat(awaitItem()).isInstanceOf<PaymentSheetScreen.VerticalMode>()
+                assertThat(awaitItem()).isInstanceOf<SelectSavedPaymentMethods>()
             }
         }
 
     @Test
-    fun `currentScreen is VerticalMode if supportedPaymentMethods is greater than 1 in Automatic`() =
+    fun `currentScreen is AddFirstPaymentMethod if supportedPaymentMethods is 2 in Automatic`() =
         runTest {
             val viewModel = createViewModel(
                 args = PAYMENT_OPTION_CONTRACT_ARGS.updateState(
@@ -537,6 +537,28 @@ internal class PaymentOptionsViewModelTest {
                     linkState = null,
                     stripeIntent = PaymentIntentFixtures.PI_WITH_PAYMENT_METHOD!!.copy(
                         paymentMethodTypes = listOf("card", "cashapp")
+                    ),
+                )
+            )
+
+            viewModel.navigationHandler.currentScreen.test {
+                assertThat(awaitItem()).isInstanceOf<AddFirstPaymentMethod>()
+            }
+        }
+
+    @Test
+    fun `currentScreen is VerticalMode if supportedPaymentMethods is 3 or more in Automatic`() =
+        runTest {
+            val viewModel = createViewModel(
+                args = PAYMENT_OPTION_CONTRACT_ARGS.updateState(
+                    config = PaymentSheetFixtures.CONFIG_CUSTOMER_WITH_GOOGLEPAY.newBuilder()
+                        .paymentMethodLayout(PaymentSheet.PaymentMethodLayout.Automatic)
+                        .build(),
+                    paymentMethods = listOf(),
+                    isGooglePayReady = false,
+                    linkState = null,
+                    stripeIntent = PaymentIntentFixtures.PI_WITH_PAYMENT_METHOD!!.copy(
+                        paymentMethodTypes = listOf("card", "cashapp", "klarna")
                     ),
                 )
             )